Changed how Gaussian treats input
Gaussian object now accepts needs parameters (list of 3 numbers). If time_array is passed as kwarg, a profile is returned. If time_array and data_array is passed as kwargs, a fit is performed. In this case, parameters can be None or it is used as initial guess for the fit.
Need to update doc string, but it's Friday evening now...